+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# unreleased
+* Add ability to/from import/export with TappedOut.net files
+
 # 0.1.0
 * Add import and export buttons in settings modal